Code cleanup (Jelle taking over)

+++ backend/menu-service/src/test/java/com/dma/menuservice/MenuRepositoryTests.java
@@ -4,20 +4,13 @@ import com.dma.menuservice.models.Menu;
 import com.dma.menuservice.models.MenuItem;
 import com.dma.menuservice.models.Restaurant;
 import com.dma.menuservice.repositories.MenuRepository;
-import org.assertj.core.util.Lists;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import org.springframework.boot.test.autoconfigure.orm.jpa.TestEntityManager;
-import org.springframework.test.context.ActiveProfiles;
-import org.springframework.test.context.junit4.SpringRunner;
 
-import java.awt.*;
-import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 import java.util.NoSuchElementException;
@@ -32,6 +25,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 @DataJpaTest
 public class MenuRepositoryTests {
+
     @Autowired
     private MenuRepository repository;
 
@@ -43,7 +37,6 @@ public class MenuRepositoryTests {
 
     /**
      * Initialize persistent data in the entity manager
-     *
      */
     @BeforeEach
     public void init() {
@@ -60,7 +53,7 @@ public class MenuRepositoryTests {
         testMenu = new Menu();
         testMenu.setName("TestMenu");
         testMenu.setRestaurantId(testRestaurant.getId());
-        testItems =  Arrays.asList(testItemOne,testItemTwo);
+        testItems = Arrays.asList(testItemOne, testItemTwo);
         testMenu.setItems(testItems);
 
         entityManager.persist(testMenu);
@@ -70,12 +63,11 @@ public class MenuRepositoryTests {
 
     /**
      * Get a single menu
-     *
      */
     @Test
     public void should_get_single_menu() {
         Long id = (Long) entityManager.getId(testMenu);
-        Menu foundMenu = repository.findById((Long) id).get();
+        Menu foundMenu = repository.findById(id).get();
 
         assertThat(foundMenu.getName()).isEqualTo("TestMenu");
         assertThat(foundMenu.getRestaurantId()).isEqualTo(1L);
@@ -85,11 +77,10 @@ public class MenuRepositoryTests {
 
     /**
      * Cannot get non-existing menu
-     *
      */
     @Test
     public void should_not_get_non_existing_menu() {
-        Assertions.assertThrows(NoSuchElementException.class,()-> {
+        Assertions.assertThrows(NoSuchElementException.class, () -> {
             Menu foundMenu = repository.findById(3L).get();
         });
 
@@ -97,7 +88,6 @@ public class MenuRepositoryTests {
 
     /**
      * Create new menu
-     *
      */
     @Test
     public void should_create_new_menu() {
@@ -112,7 +102,7 @@ public class MenuRepositoryTests {
         repository.save(newMenu);
 
         Long id = (Long) entityManager.getId(newMenu);
-        Menu foundMenu = repository.findById((Long) id).get();
+        Menu foundMenu = repository.findById(id).get();
 
         assertThat(foundMenu.getName()).isEqualTo(newMenu.getName());
         assertThat(foundMenu.getItems().size()).isEqualTo(1);
@@ -121,12 +111,11 @@ public class MenuRepositoryTests {
 
     /**
      * Update menu
-     *
      */
     @Test
     public void should_update_menu() {
         Long id = (Long) entityManager.getId(testMenu);
-        Menu foundMenu = repository.findById((Long) id).get();
+        Menu foundMenu = repository.findById(id).get();
 
         MenuItem testItemThree = new MenuItem();
         testItemThree.setName("Fanta 330ml");
@@ -134,7 +123,7 @@ public class MenuRepositoryTests {
         foundMenu.setItems(Arrays.asList(testItemThree));
         foundMenu.setName("UpdatedTestRestaurant");
 
-        Menu updatedMenu = entityManager.find(Menu.class,foundMenu.getId());
+        Menu updatedMenu = entityManager.find(Menu.class, foundMenu.getId());
 
         assertThat(updatedMenu.getItems().contains(testItemThree)).isTrue();
         assertThat(updatedMenu.getItems().size()).isNotEqualTo(2);
@@ -143,12 +132,11 @@ public class MenuRepositoryTests {
 
     /**
      * Delete menu
-     *
      */
     @Test
     public void should_delete_menu() {
         Long id = (Long) entityManager.getId(testMenu);
-        Menu foundMenu = repository.findById((Long) id).get();
+        Menu foundMenu = repository.findById(id).get();
         repository.delete(foundMenu);
 
         assertThat(repository.findAll()).hasSize(0);
